Renamed all functions and fixed indexes of all Extensions controllers
